Final Product Changes
This is the update that I made that completed the program. Complete in
the sense that it's now a working app.

MAJOR CHANGES:

- Completed functions for 1.5, 2 and 2.5 overtime hours.

- Removed the PrintTotal() function and had the running total display
below the menu every time you return to the menu. Seemed redundant to
have to choose to see the running total from a menu.

MINOR CHANGES:

These are mainly cosmetic changes.

- Moved the def <function> code snippets to the top of the file. This is
just for consistency. Functions should all be defined at the top before
the main code executes to avoid confusion.

- In the beginning of the program where the user attributes (OT rates)
are collected, I was calculating the pay rates then rounding them on
separate lines. I've now moved those calculations onto single lines for
each rate.

The program works! Yay!
Next up will be storing the total and indexing the user attributes file.
I'm currently relying on pulling values based on the line they're on.
This isn't the most effective and accurate way to handle this (but works
for now!).